Fix ViewModeFlags; WorkspaceController owns session-id remove
ViewMode is nested in ImageView — ViewModeFlags takes int mode and casts
to ImageView::ViewMode in the .cpp.

Move removeWorkspaceSessionId and helpers (collect/destroy session-id items,
prune binds and path order, restore viewport) onto WorkspaceController.
ImageView keeps thin routers.

+++ src/imageview_modes.cpp
@@ -211,7 +211,7 @@ void ImageView::setViewMode(ViewMode mode)
 
 void ImageView::applyItemModeFlags(ImageItem *item)
 {
-    ViewModeFlags::applyToItem(item, m_viewMode);
+    ViewModeFlags::applyToItem(item, static_cast<int>(m_viewMode));
 }
 
 void ImageView::setLayoutMode(LayoutMode mode)
